Conflicts:
	Game/CreatePlayerGUI.java
	Game/StartMenuGUI.java

+++ Board/Animate.java
@@ -1,10 +1,15 @@
-/**
- * This interface is used for sprite animation
+/*
+ * To change this template, choose Tools | Templates
+ * and open the template in the editor.
  */
-
 package Board;
 
+
+/**
+ *
+ * @author ooo524
+ */
 public interface Animate {
-    void move( double x, double y);
+    void move( int x, int y);
     void changeConfiguration();
 }
